financial-scam-detector/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML & NLP - Updated for compatibility
transformers>=4.30.0
torch>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.3
pandas>=2.1.0

# Text processing
nltk>=3.8.0
regex>=2023.10.0

# Domain checking
python-whois>=0.8.0
dnspython>=2.4.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
